@@ -1,12 +1,32 @@
-module github.com/infinityworks/github-exporter
+module github.com/githubexporter/github-exporter
 
-go 1.14
+go 1.22
 
 require (
-	github.com/fatih/structs v1.1.0
+	github.com/bradleyfalzon/ghinstallation/v2 v2.11.0
 	github.com/infinityworks/go-common v0.0.0-20170820165359-7f20a140fd37
-	github.com/prometheus/client_golang v0.9.2
-	github.com/sirupsen/logrus v1.4.2
+	github.com/prometheus/client_golang v1.20.4
+	github.com/sirupsen/logrus v1.9.3
 	github.com/steinfletcher/apitest v1.3.8
 	github.com/tomnomnom/linkheader v0.0.0-20180905144013-02ca5825eb80
+)
+
+require (
+	github.com/beorn7/perks v1.0.1 // indirect
+	github.com/cespare/xxhash/v2 v2.3.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/golang-jwt/jwt/v4 v4.5.0 // indirect
+	github.com/google/go-github/v62 v62.0.0 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
+	github.com/klauspost/compress v1.17.10 // indirect
+	github.com/kr/text v0.2.0 // indirect
+	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/prometheus/client_model v0.6.1 // indirect
+	github.com/prometheus/common v0.59.1 // indirect
+	github.com/prometheus/procfs v0.15.1 // indirect
+	github.com/stretchr/testify v1.9.0 // indirect
+	golang.org/x/sys v0.25.0 // indirect
+	google.golang.org/protobuf v1.34.2 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
