@@ -2,7 +2,7 @@ package exporter
 
 import (
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	neturl "net/url"
 	"strconv"
@@ -40,7 +40,7 @@ func asyncHTTPGets(targets []string, token string) ([]*Response, error) {
 		case r := <-ch:
 			if r.err != nil {
 				log.Errorf("Error scraping API, Error: %v", r.err)
-				break
+				return nil, r.err
 			}
 			responses = append(responses, r)
 
@@ -111,14 +111,14 @@ func getResponse(url string, token string, ch chan<- *Response) error {
 	defer resp.Body.Close()
 
 	// Read the body to a byte array so it can be used elsewhere
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return fmt.Errorf("Error converting body to byte array: %v", err)
 	}
 
 	// Triggers if a user specifies an invalid or not visible repository
 	if resp.StatusCode == 404 {
-		return fmt.Errorf("Error: Received 404 status from Github API, ensure the repsository URL is correct. If it's a privare repository, also check the oauth token is correct")
+		return fmt.Errorf("Error: Received 404 status from Github API, ensure the repository URL is correct. If it's a private repository, also check the oauth token is correct")
 	}
 
 	ch <- &Response{url, resp, body, err}
